requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
et-xmlfile==1.1.0
executing==2.1.0
fake-useragent==1.5.1
fake_user_agent==2.1.8
filelock==3.14.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
httpx-html==0.11.0.dev0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
incremental==22.10.0
ipykernel==6.29.5
ipython==8.28.0
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaitaistruct==0.10
lxml==5.2.1
lxml_html_clean==0.1.1
matplotlib-inline==0.1.7
multidict==6.0.5
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parse==1.20.1
parsel==1.9.1
parso==0.8.4
platformdirs==4.2.2
playwright==1.45.1
prompt_toolkit==3.0.48
Protego==0.3.1
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pydivert==2.1.0
pyee==11.1.0
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.2.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
selectolax==0.3.21
selenium==4.23.1
selenium-stealth==1.0.6
selenium-wire==5.1.0
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
stack-data==0.6.3
tldextract==5.1.2
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
w3lib==2.1.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
zope.interface==6.4.post2
zstandard==0.23.0
